Move NotebookParser to pipeline-editor
The utility class NotebookParser is only used in pipeline-editor
