+++ packages/application/src/index.ts
@@ -16,6 +16,5 @@
 
 import '../style/index.css';
 
-export * from './parsing';
 export * from './services';
 export * from './requests';
