+++ packages/application/src/services.tsx
@@ -17,8 +17,7 @@
 import { Dialog, showDialog } from '@jupyterlab/apputils';
 import * as React from 'react';
 
-import { IDictionary } from './parsing';
-import { RequestHandler } from './requests';
+import { IDictionary, RequestHandler } from './requests';
 
 const ELYRA_SCHEMA_API_ENDPOINT = 'elyra/schema/';
 const ELYRA_METADATA_API_ENDPOINT = 'elyra/metadata/';
