+++ packages/pipeline-editor/src/NotebookParser.ts
@@ -15,13 +15,6 @@
  */
 
 /**
- * An interface for typing json dictionaries in typescript
- */
-export interface IDictionary<T> {
-  [key: string]: T;
-}
-
-/**
  * A utilities class for parsing notebook files.
  */
 export class NotebookParser {
